+++ fridge/src/main/java/at/jku/ce/fridge/service/IFridgeService.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IFridgeService extends JpaRepository<Fridge,Integer> {
+public interface IFridgeService extends JpaRepository<Fridge,Long> {
 }
